Update README.rst
Dynamic fields didn't work. Solution: Adding the filter to the handler instead of adding it to the logger.

In my scenario (flask app), I didn't get the dynamic fields to work the way you described it in this read-me. However, if I add the filter to the handler everything works perfectly :)